openai_agents_lecture/requirements.txt
openai-agents
python-dotenv